requirements.txt
pygal
flask
gevent
python-dateutil
